repository for general activity

+++ src/main/java/backend/siptis/model/repository/records/GeneralActivityRepository.java
@@ -3,5 +3,5 @@ package backend.siptis.model.repository.records;
 import backend.siptis.model.entity.records.GeneralActivity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GeneralActivityRepository extends JpaRepository<GeneralActivity, Integer> {
+public interface GeneralActivityRepository extends JpaRepository<GeneralActivity, Long> {
 }
